WIP for fixing unhandled error in supervisor assigns volunteers
