+++ app/controllers/supervisor_volunteers_controller.rb
@@ -3,6 +3,7 @@ class SupervisorVolunteersController < ApplicationController
   before_action :must_be_admin_or_supervisor, only: :unassign
 
   def create
+  
     supervisor_volunteer = supervisor_volunteer_parent.supervisor_volunteers.find_or_create_by!(supervisor_volunteer_params)
     supervisor_volunteer.is_active = true unless supervisor_volunteer.is_active?
     supervisor_volunteer.save
